+++ packages/hightide-native/eslint.config.js
@@ -27,7 +27,7 @@ export default [
     },
   },
   {
-    ignores: ['node_modules/**'],
+    ignores: ['dist/**', 'node_modules/**'],
   },
   ...config.recommended,
   {