+++ packages/hightide/src/global-contexts/theme/ThemeProvider.tsx
@@ -12,6 +12,7 @@ import type {
 } from '@helpwave/hightide-utils/context/theme'
 import { useCreateThemeConfig } from '@helpwave/hightide-utils/context/theme'
 import { ThemeContext } from './ThemeContext'
+import { SafeGlobals } from '../../utils/safeGlobals'
 
 export type ThemeInformation = Omit<ThemeInformationBase<unknown>, 'theme'> & { theme?: unknown }
 export type SupportedThemesConfig = Record<SystemTheme, ThemeInformation> & Record<string, ThemeInformation>
@@ -41,7 +42,7 @@ export const ThemeProvider = ({
       },
       dark: {
         ...resolvedSupportedThemes.dark,
-        theme: resolvedSupportedThemes.light.theme ?? undefined,
+        theme: resolvedSupportedThemes.dark.theme ?? undefined,
       }
     }
     for(const themeMode in resolvedSupportedThemes) {
@@ -64,7 +65,7 @@ export const ThemeProvider = ({
   })
 
   useEffect(() => {
-    document.documentElement.setAttribute('data-theme', value.themeMode)
+    SafeGlobals.document('ThemeProvider')?.documentElement.setAttribute('data-theme', value.themeMode)
   }, [value.themeMode])
 
   return (